# sso-service/docs/swagger.yaml
basePath: /
definitions:
...
  participant_dto.ParticipantResponseDTO:
    properties:
      citizenship:
        type: string
      class_number:
        type: string
      disability:
        type: string
      id:
        type: string
      school_id:
        type: string
      user_id:
        type: string
    type: object
  participant_dto.UpdateParticipantRequestDTO:
    properties:
      citizenship:
        type: string
      class_number:
        type: string
      disability:
        type: string
      school_id:
        type: string
    type: object
...
  response.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      metadata: {}
      status:
        type: string
      status_code:
        type: integer
    type: object
  school_dto.CreateSchoolRequestDTO:
    properties:
      name:
        type: string
      region:
        type: string
    type: object
  school_dto.SchoolResponseDTO:
    properties:
      id:
        type: string
      name:
        type: string
      region:
        type: integer
    type: object
  school_dto.UpdateSchoolRequestDTO:
    properties:
      name:
        type: string
      region:
        type: string
    type: object
...
host: localhost:8181
info:
  contact: {}
  description: Документация к микросервису авторизации
  title: SSO Service API
  version: "1.0"
paths:
...
  /api/participants:
    get:
      description: Получение всех учеников
      parameters:
      - description: Номер страницы
        in: query
        name: page
        type: integer
      - description: Ограничение на количество записей
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/participant_dto.ParticipantResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - participants
  /api/participants/{id}:
    get:
      description: Получение данных о ученике по id
      parameters:
      - description: id ученика
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - participants
    put:
      consumes:
      - application/json
      description: Обновление данных об ученике
      parameters:
      - description: Новые данные для обновления
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/participant_dto.UpdateParticipantRequestDTO'
      - description: id пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - participants
  /api/participants/byuser/{id}:
    get:
      description: Получение данных о ученике по user id
      parameters:
      - description: id пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - participants
  /api/participants/count:
    get:
      description: Получение количества участников
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - participants
  /api/schools:
    get:
      description: Получение всех школ
      parameters:
      - description: Номер страницы
        in: query
        name: page
        type: integer
      - description: Ограничение на количество записей
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/school_dto.SchoolResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
    post:
      consumes:
      - application/json
      description: Создание школы
      parameters:
      - description: Данные для создания школы
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/school_dto.CreateSchoolRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
  /api/schools/{id}:
    get:
      description: Получение школы по id
      parameters:
      - description: id школы
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/school_dto.SchoolResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
    put:
      consumes:
      - application/json
      description: Обновление школы
      parameters:
      - description: Данные для обновления школы
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/school_dto.UpdateSchoolRequestDTO'
      - description: id пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
  /api/schools/count:
    get:
      description: Получение количества школ
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      tags:
      - schools
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"